now single precision numpy arrays are set at the moment of its creation. As a result we don't need to recast from np.float64 to np.float32 every time we pass these arrays to the extension modules.

+++ tupan/particles/blackhole.py
@@ -9,6 +9,7 @@ from __future__ import print_function
 import numpy as np
 from .body import Bodies
 from ..lib.utils.timing import decallmethods, timings
+from ..lib.utils.dtype import *
 
 
 __all__ = ["Blackholes"]
@@ -20,10 +21,10 @@ class Blackholes(Bodies):
 
     """
     dtype = Bodies.dtype + [
-                            ("sx", np.float64),
-                            ("sy", np.float64),
-                            ("sz", np.float64),
-                            ("radius", np.float64),
+                            ("sx", REAL),
+                            ("sy", REAL),
+                            ("sz", REAL),
+                            ("radius", REAL),
                            ]
